ecommerce/requirements.txt
Django==5.1.7
django_phonenumber_field==8.0.0
